+++ internal/metrics/testdata/execute_happy.txt
