+++ internal/metrics/testdata/execute_nometrics.txt
